Started making test case for Issue #36
