+++ Tests/Regression/Reporter/pip_requirements.txt
@@ -1,5 +1,6 @@
+# This is required because of: https://github.com/langchain-ai/langchain/issues/26026 and https://github.com/numpy/numpy/issues/24318
 numpy>2; python_version > '3.12' and platform_system == 'Windows'
-PyMuPDF>=1.23.1; python_version > '3.12' and platform_system == 'Windows'
+PyMuPDF>=1.23.1; python_version > '3.12' and platform_system == 'Windows' # require numpy > 2
 robotframework
 robotframework-doctestlibrary
 # pyscreeze 0.1.29 doesn't seem to work with pyautogui 0.9.53
@@ -16,3 +17,4 @@ python-docx
 openpyxl
 openpyxl_image_loader
 pandas
+pyyaml
